--- client/src/pages/Dashboard.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Dashboard = () => {
-  return (
-    <div>Dashboard</div>
-  )
-}
-
-export default Dashboard
